Refactors request handling for PG and User
Updates the PGController and UserController to handle request bodies as Maps instead of directly binding to entity objects. This change provides more flexibility in request parameter handling. It also adds validation to ensure users exist before creating a PG and handles null values for URL fields in User and PG entities.

Updated PG creation logic to add the user who created the PG as a member of that pg and then as admin for that pg